split edit panel
