+++ src/constants/pdf_tool.ts
@@ -1,0 +1,4 @@
+export enum PdfToolViewMode {
+  EDIT = "edit",
+  PREVIEW = "preview",
+}
